Only delete files if check = T

+++ R/roclet-rd.R
@@ -165,13 +165,16 @@ roc_output.had <- function(roclet, results, base_path, options = list(),
   paths <- file.path(man, names(results))
   mapply(write_if_different, paths, contents, MoreArgs = list(check = check))
 
-  # Automatically delete any files in man directory that were generated
-  # by roxygen in the past, but weren't generated in this sweep.
-  old_paths <- setdiff(dir(man, full.names = TRUE), paths)
-  old_roxygen <- Filter(made_by_roxygen, old_paths)
-  if (length(old_roxygen) > 0) {
-    cat(paste0("Deleting ", basename(old_roxygen), collapse = "\n"), "\n", sep = "")
-    unlink(old_roxygen)
+  if (check) {
+    # Automatically delete any files in man directory that were generated
+    # by roxygen in the past, but weren't generated in this sweep.
+
+    old_paths <- setdiff(dir(man, full.names = TRUE), paths)
+    old_roxygen <- Filter(made_by_roxygen, old_paths)
+    if (length(old_roxygen) > 0) {
+      cat(paste0("Deleting ", basename(old_roxygen), collapse = "\n"), "\n", sep = "")
+      unlink(old_roxygen)
+    }
   }
 
   paths
